android_menu_scraper/requirements.txt
appium-python-client==3.2.0
pillow>=10.0.0
pytesseract>=0.3.10
lxml>=5.1.0
